backend/requirements.txt
pymango
datetime
langchain
pyttsx3
dotenv
google.generativeai
python-pptx
asyncio
edge_tts
pydub
langchain-community
pypdf
python-dotenv
groq
gtts
langchain_groq
langchain_ollama
chromadb
#download ollama 